leave base class for later, fix some typos, small regression tested and looks good
